fix(clippy): Issue 701 R3b slice 3 - all-features warning surface 67->13, deny the 15 settled lints
cargo clippy --fix (15 edits, 8 files) + 34 hand-rewritten needless_range_loop
sites (8 triangular symmetric fills via flat_map tuple iteration preserving the
exact LCG draw order; 10 trivial iter/enumerate/zip rewrites). Fix-introduced
findings cleaned in the same pass: iter_cloned_collect x2 (.to_vec()) +
unused_parens x1 (clippy's own identity_op fix emitted them).

Residual 13 = documented judgement classes only (unused_variables x4,
dead_code x3, non_snake_case x2, too_many_arguments x2,
assertions_on_constants x2) - left untouched per the R3a decision.

full_gate.sh GATE_ARGS + AGENTS.md parity: -D the 15 zero-residual lints
(needless_range_loop, map_clone, iter_cloned_collect, identity_op,
unused_parens, bool_comparison, manual_is_multiple_of, collapsible_if,
map_all_any_identity, unnecessary_cast, manual_repeat_n, question_mark,
empty_line_after_outer_attr, unusual_byte_groupings, unused_mut).
FULL_GATE_LOG run: PASS, 0 errors, 13 warnings, 32 units.

Validation: katgpt-core 1992/0/7i (count exact), pruners 126/0, speculative
305/0, types 132/0, attn+flashmemory_sparse 174/0, root lib 201/0, issue_698
t6/t8 3/0, issue_699 15/0, engram_tripwire 4/0, bench_582 5/0, bench_684
run VERDICT PASS.

+++ tests/issue_698_t8_gate_probe.rs
@@ -233,7 +233,7 @@ fn spearman(xs: &[f32], ys: &[f32]) -> f32 {
     rank_average(xs)
         .iter()
         .zip(rank_average(ys).iter())
-        .map(|(&a, &b)| (a as f64, b as f64))
+        .map(|(&a, &b)| (a, b))
         .collect::<Vec<_>>()
         .pipe(|pairs| pearson_f64(&pairs)) as f32
 }
